properly structure rfkill util
